+++ internal/checks/rule_dependency.go
@@ -210,10 +210,7 @@ func (c RuleDependencyCheck) checkCrossGroupDependency(entry *discovery.Entry, e
 				if isStaticRecordingRule(other) {
 					continue
 				}
-				if other.Group == nil {
-					continue
-				}
-				if entry.Group != nil && entry.Group.Name.Value == other.Group.Name.Value {
+				if entry.Group != nil && other.Group != nil && entry.Group.Name.Value == other.Group.Name.Value {
 					continue
 				}
 				vsPos := vs.PositionRange()